+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
   for Javascript/Typescript and Java.
 - Added matching of annotation patterns for Java (@SomeAnnot(...)) in any
   context
+- The `metavariable-comparison` operator
 
 ## [0.27.0](https://github.com/returntocorp/semgrep/releases/tag/v0.27.0) - 2020-10-06